Added everything for playtest
First made the item spawner where items spawn in the air and then come falling down.

Was hard to see at first so i then made a script that would shoot a raycast down to the floor and place a square/circle that is supposed to represent where the item is going so you can set yourself up with it

Then added catch and reward system with points

Added cinematic camera to get players to lock in and try their best to balance the items they have on their tray.

A lot of work for one day im beat.